Restructure NewDisplayAsset and their related classes in new_display_asset.py
