+++ fortnite_api/new_display_asset.py
@@ -24,7 +24,7 @@ SOFTWARE.
 
 from __future__ import annotations
 
-from typing import Any, Dict, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
 from .abc import Hashable, ReconstructAble
 from .asset import Asset
@@ -33,10 +33,40 @@ from .http import HTTPClientT
 from .utils import get_with_fallback
 
 __all__: Tuple[str, ...] = (
+    "RenderImage",
     "MaterialInstanceImages",
     "MaterialInstanceColors",
     "MaterialInstance",
+    "NewDisplayAsset",
 )
+
+
+class RenderImage(ReconstructAble[Dict[str, Any], HTTPClientT]):
+    """
+    .. attributetable:: fortnite_api.RenderImage
+
+    Represents a render image for a shop entry. A render image is an image
+    used to visually represent a cosmetic item in the shop. This class inherits
+    from :class:`~fortnite_api.ReconstructAble`.
+
+    Attributes
+    ----------
+    product_tag: :class:`fortnite_api.ProductTag`
+        The product tag of the render image.
+    file_name: :class:`str`
+        The internal file name of the rendered image. Refers to the name within the game files and **not** the :attr`image`. An example of this is ``T-Featured-Pickaxes-SleepyTimePickaxe``.
+    image: :class:`fortnite_api.Asset`
+        The image of the render image.
+    """
+
+    __slots__: Tuple[str, ...] = ("product_tag", "file_name", "image")
+
+    def __init__(self, *, data: Dict[str, Any], http: HTTPClientT) -> None:
+        super().__init__(data=data, http=http)
+
+        self.product_tag: ProductTag = ProductTag._from_str(data["productTag"])
+        self.file_name: str = data["fileName"]
+        self.image: Asset[HTTPClientT] = Asset(url=data["image"], http=http)
 
 
 class MaterialInstanceImages(Dict[str, Asset[HTTPClientT]]):
@@ -169,3 +199,38 @@ class MaterialInstance(Hashable, ReconstructAble[Dict[str, Any], HTTPClientT]):
         self.scalings: Dict[str, Any] = get_with_fallback(data, "scalings", dict)
 
         self.flags: Dict[str, Any] = get_with_fallback(data, "flags", dict)  # This is always None at this time.
+
+
+class NewDisplayAsset(Hashable, ReconstructAble[Dict[str, Any], HTTPClientT]):
+    """
+    .. attributetable:: fortnite_api.NewDisplayAsset
+
+    Represents a new display asset for a shop entry. A display asset is an asset that is
+    used to visually represent a cosmetic item in the shop. This class inherits
+    from :class:`~fortnite_api.ReconstructAble`.
+
+    Attributes
+    ----------
+    id: :class:`str`
+        The ID of the display asset.
+    cosmetic_id: Optional[:class:`str`]
+        The ID of the cosmetic item associated with the display asset, if any.
+    material_instances: List[:class:`fortnite_api.MaterialInstance`]
+        A list of material instances used by the display asset.
+    render_images: List[:class:`fortnite_api.RenderImage`]
+        A list of render images used by the display asset.
+    """
+
+    __slots__: Tuple[str, ...] = ("id", "cosmetic_id", "material_instances", "render_images")
+
+    def __init__(self, *, data: Dict[str, Any], http: HTTPClientT) -> None:
+        super().__init__(data=data, http=http)
+
+        self.id: str = data["id"]
+        self.cosmetic_id: Optional[str] = data.get("cosmeticId")
+        self.material_instances: List[MaterialInstance[HTTPClientT]] = [
+            MaterialInstance(data=instance, http=http) for instance in get_with_fallback(data, "materialInstances", list)
+        ]
+        self.render_images: List[RenderImage[HTTPClientT]] = [
+            RenderImage(data=instance, http=http) for instance in get_with_fallback(data, "renderImages", list)
+        ]
